final version

+++ include/bfile.h
@@ -2,13 +2,15 @@
 #define __BFILE_H_
 #define N 4096
 
+#include <unistd.h>
+
 struct bfile;
 typedef struct bfile bfile;
 
 bfile * bOpen (const char * path, char mode);
 int bClose(bfile * bf);
-int bWrite(void * p, int size, int nb_element, bfile * bf);
-int bRead(void * p, int size, int nb_element, bfile * bf);
-int bFlush(bfile * bf);
-int bFill(bfile * bf);
+ssize_t bWrite(void * p, int size, int nb_element, bfile * bf);
+ssize_t bRead(void * p, int size, int nb_element, bfile * bf);
+ssize_t bFlush(bfile * bf);
+ssize_t bFill(bfile * bf);
 #endif
